text_autoencoder_pretrain/requirements.txt
nltk
docopt
numpy
scipy
tqdm
docopt
pytorch
torchvision
